cassandra buffered rate executor: separate beans for read and write operations
