--- dao/src/main/java/org/thingsboard/server/dao/nosql/CassandraBufferedRateExecutor.java
@@ -1,140 +1,0 @@
-/**
- * Copyright © 2016-2021 The Thingsboard Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.thingsboard.server.dao.nosql;
-
-import com.google.common.util.concurrent.ListenableFuture;
-import com.google.common.util.concurrent.SettableFuture;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Component;
-import org.thingsboard.server.common.data.id.TenantId;
-import org.thingsboard.server.common.stats.DefaultCounter;
-import org.thingsboard.server.common.stats.StatsCounter;
-import org.thingsboard.server.common.stats.StatsFactory;
-import org.thingsboard.server.dao.entity.EntityService;
-import org.thingsboard.server.dao.util.AbstractBufferedRateExecutor;
-import org.thingsboard.server.dao.util.AsyncTaskContext;
-import org.thingsboard.server.dao.util.NoSqlAnyDao;
-
-import javax.annotation.PreDestroy;
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * Created by ashvayka on 24.10.18.
- */
-@Component
-@Slf4j
-@NoSqlAnyDao
-public class CassandraBufferedRateExecutor extends AbstractBufferedRateExecutor<CassandraStatementTask, TbResultSetFuture, TbResultSet> {
-
-    @Autowired
-    private EntityService entityService;
-    private Map<TenantId, String> tenantNamesCache = new HashMap<>();
-
-    private boolean printTenantNames;
-
-    public CassandraBufferedRateExecutor(
-            @Value("${cassandra.query.buffer_size}") int queueLimit,
-            @Value("${cassandra.query.concurrent_limit}") int concurrencyLimit,
-            @Value("${cassandra.query.permit_max_wait_time}") long maxWaitTime,
-            @Value("${cassandra.query.dispatcher_threads:2}") int dispatcherThreads,
-            @Value("${cassandra.query.callback_threads:4}") int callbackThreads,
-            @Value("${cassandra.query.poll_ms:50}") long pollMs,
-            @Value("${cassandra.query.tenant_rate_limits.enabled}") boolean tenantRateLimitsEnabled,
-            @Value("${cassandra.query.tenant_rate_limits.configuration}") String tenantRateLimitsConfiguration,
-            @Value("${cassandra.query.tenant_rate_limits.print_tenant_names}") boolean printTenantNames,
-            @Value("${cassandra.query.print_queries_freq:0}") int printQueriesFreq,
-            @Autowired StatsFactory statsFactory) {
-        super(queueLimit, concurrencyLimit, maxWaitTime, dispatcherThreads, callbackThreads, pollMs, tenantRateLimitsEnabled, tenantRateLimitsConfiguration, printQueriesFreq, statsFactory);
-        this.printTenantNames = printTenantNames;
-    }
-
-    @Scheduled(fixedDelayString = "${cassandra.query.rate_limit_print_interval_ms}")
-    public void printStats() {
-        int queueSize = getQueueSize();
-        int rateLimitedTenantsCount = (int) stats.getRateLimitedTenants().values().stream()
-                .filter(defaultCounter -> defaultCounter.get() > 0)
-                .count();
-
-        if (queueSize > 0
-                || rateLimitedTenantsCount > 0
-                || concurrencyLevel.get() > 0
-                || stats.getStatsCounters().stream().anyMatch(counter -> counter.get() > 0)
-        ) {
-            StringBuilder statsBuilder = new StringBuilder();
-
-            statsBuilder.append("queueSize").append(" = [").append(queueSize).append("] ");
-            stats.getStatsCounters().forEach(counter -> {
-                statsBuilder.append(counter.getName()).append(" = [").append(counter.get()).append("] ");
-            });
-            statsBuilder.append("totalRateLimitedTenants").append(" = [").append(rateLimitedTenantsCount).append("] ");
-            statsBuilder.append(CONCURRENCY_LEVEL).append(" = [").append(concurrencyLevel.get()).append("] ");
-
-            stats.getStatsCounters().forEach(StatsCounter::clear);
-            log.info("Permits {}", statsBuilder);
-        }
-
-        stats.getRateLimitedTenants().entrySet().stream()
-                .filter(entry -> entry.getValue().get() > 0)
-                .forEach(entry -> {
-                    TenantId tenantId = entry.getKey();
-                    DefaultCounter counter = entry.getValue();
-                    int rateLimitedRequests = counter.get();
-                    counter.clear();
-                    if (printTenantNames) {
-                        String name = tenantNamesCache.computeIfAbsent(tenantId, tId -> {
-                            try {
-                                return entityService.fetchEntityNameAsync(TenantId.SYS_TENANT_ID, tenantId).get();
-                            } catch (Exception e) {
-                                log.error("[{}] Failed to get tenant name", tenantId, e);
-                                return "N/A";
-                            }
-                        });
-                        log.info("[{}][{}] Rate limited requests: {}", tenantId, name, rateLimitedRequests);
-                    } else {
-                        log.info("[{}] Rate limited requests: {}", tenantId, rateLimitedRequests);
-                    }
-                });
-    }
-
-    @PreDestroy
-    public void stop() {
-        super.stop();
-    }
-
-    @Override
-    protected SettableFuture<TbResultSet> create() {
-        return SettableFuture.create();
-    }
-
-    @Override
-    protected TbResultSetFuture wrap(CassandraStatementTask task, SettableFuture<TbResultSet> future) {
-        return new TbResultSetFuture(future);
-    }
-
-    @Override
-    protected ListenableFuture<TbResultSet> execute(AsyncTaskContext<CassandraStatementTask, TbResultSet> taskCtx) {
-        CassandraStatementTask task = taskCtx.getTask();
-        return task.executeAsync(
-                statement ->
-                    this.submit(new CassandraStatementTask(task.getTenantId(), task.getSession(), statement))
-        );
-    }
-
-}
