+++ dao/src/main/java/org/thingsboard/server/dao/util/AbstractBufferedRateExecutor.java
@@ -31,12 +31,17 @@ import lombok.extern.slf4j.Slf4j;
 import org.thingsboard.common.util.ThingsBoardExecutors;
 import org.thingsboard.common.util.ThingsBoardThreadFactory;
 import org.thingsboard.server.common.data.id.TenantId;
+import org.thingsboard.server.common.msg.tools.TbRateLimits;
+import org.thingsboard.server.common.stats.DefaultCounter;
+import org.thingsboard.server.common.stats.StatsCounter;
 import org.thingsboard.server.common.stats.StatsFactory;
 import org.thingsboard.server.common.stats.StatsType;
-import org.thingsboard.server.common.msg.tools.TbRateLimits;
+import org.thingsboard.server.dao.entity.EntityService;
 import org.thingsboard.server.dao.nosql.CassandraStatementTask;
 
 import javax.annotation.Nullable;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.ConcurrentHashMap;
@@ -75,21 +80,31 @@ public abstract class AbstractBufferedRateExecutor<T extends AsyncTask, F extend
     protected final AtomicInteger concurrencyLevel;
     protected final BufferedRateExecutorStats stats;
 
+
+    private final EntityService entityService;
+    private final Map<TenantId, String> tenantNamesCache = new HashMap<>();
+
+    private final boolean printTenantNames;
+
     public AbstractBufferedRateExecutor(int queueLimit, int concurrencyLimit, long maxWaitTime, int dispatcherThreads, int callbackThreads, long pollMs,
-                                        boolean perTenantLimitsEnabled, String perTenantLimitsConfiguration, int printQueriesFreq, StatsFactory statsFactory) {
+                                        boolean perTenantLimitsEnabled, String perTenantLimitsConfiguration, int printQueriesFreq, StatsFactory statsFactory,
+                                        EntityService entityService, boolean printTenantNames) {
         this.maxWaitTime = maxWaitTime;
         this.pollMs = pollMs;
         this.concurrencyLimit = concurrencyLimit;
         this.printQueriesFreq = printQueriesFreq;
         this.queue = new LinkedBlockingDeque<>(queueLimit);
-        this.dispatcherExecutor = Executors.newFixedThreadPool(dispatcherThreads, ThingsBoardThreadFactory.forName("nosql-dispatcher"));
-        this.callbackExecutor = ThingsBoardExecutors.newWorkStealingPool(callbackThreads, getClass());
-        this.timeoutExecutor = Executors.newSingleThreadScheduledExecutor(ThingsBoardThreadFactory.forName("nosql-timeout"));
+        this.dispatcherExecutor = Executors.newFixedThreadPool(dispatcherThreads, ThingsBoardThreadFactory.forName("nosql-" + getBufferName() + "-dispatcher"));
+        this.callbackExecutor = ThingsBoardExecutors.newWorkStealingPool(callbackThreads, "nosql-" + getBufferName() + "-callback");
+        this.timeoutExecutor = Executors.newSingleThreadScheduledExecutor(ThingsBoardThreadFactory.forName("nosql-" + getBufferName() + "-timeout"));
         this.perTenantLimitsEnabled = perTenantLimitsEnabled;
         this.perTenantLimitsConfiguration = perTenantLimitsConfiguration;
         this.stats = new BufferedRateExecutorStats(statsFactory);
-        String concurrencyLevelKey = StatsType.RATE_EXECUTOR.getName() + "." + CONCURRENCY_LEVEL;
+        String concurrencyLevelKey = StatsType.RATE_EXECUTOR.getName() + "." + CONCURRENCY_LEVEL + getBufferName(); //metric name may change with buffer name suffix
         this.concurrencyLevel = statsFactory.createGauge(concurrencyLevelKey, new AtomicInteger(0));
+
+        this.entityService = entityService;
+        this.printTenantNames = printTenantNames;
 
         for (int i = 0; i < dispatcherThreads; i++) {
             dispatcherExecutor.submit(this::dispatch);
@@ -143,6 +158,8 @@ public abstract class AbstractBufferedRateExecutor<T extends AsyncTask, F extend
     protected abstract F wrap(T task, SettableFuture<V> future);
 
     protected abstract ListenableFuture<V> execute(AsyncTaskContext<T, V> taskCtx);
+
+    public abstract String getBufferName();
 
     private void dispatch() {
         log.info("Buffered rate executor thread started");
@@ -263,5 +280,52 @@ public abstract class AbstractBufferedRateExecutor<T extends AsyncTask, F extend
 
     protected int getQueueSize() {
         return queue.size();
+    }
+
+    public void printStats() {
+        int queueSize = getQueueSize();
+        int rateLimitedTenantsCount = (int) stats.getRateLimitedTenants().values().stream()
+                .filter(defaultCounter -> defaultCounter.get() > 0)
+                .count();
+
+        if (queueSize > 0
+                || rateLimitedTenantsCount > 0
+                || concurrencyLevel.get() > 0
+                || stats.getStatsCounters().stream().anyMatch(counter -> counter.get() > 0)
+        ) {
+            StringBuilder statsBuilder = new StringBuilder();
+
+            statsBuilder.append("queueSize").append(" = [").append(queueSize).append("] ");
+            stats.getStatsCounters().forEach(counter -> {
+                statsBuilder.append(counter.getName()).append(" = [").append(counter.get()).append("] ");
+            });
+            statsBuilder.append("totalRateLimitedTenants").append(" = [").append(rateLimitedTenantsCount).append("] ");
+            statsBuilder.append(CONCURRENCY_LEVEL).append(" = [").append(concurrencyLevel.get()).append("] ");
+
+            stats.getStatsCounters().forEach(StatsCounter::clear);
+            log.info("Permits {}", statsBuilder);
+        }
+
+        stats.getRateLimitedTenants().entrySet().stream()
+                .filter(entry -> entry.getValue().get() > 0)
+                .forEach(entry -> {
+                    TenantId tenantId = entry.getKey();
+                    DefaultCounter counter = entry.getValue();
+                    int rateLimitedRequests = counter.get();
+                    counter.clear();
+                    if (printTenantNames) {
+                        String name = tenantNamesCache.computeIfAbsent(tenantId, tId -> {
+                            try {
+                                return entityService.fetchEntityNameAsync(TenantId.SYS_TENANT_ID, tenantId).get();
+                            } catch (Exception e) {
+                                log.error("[{}] Failed to get tenant name", tenantId, e);
+                                return "N/A";
+                            }
+                        });
+                        log.info("[{}][{}] Rate limited requests: {}", tenantId, name, rateLimitedRequests);
+                    } else {
+                        log.info("[{}] Rate limited requests: {}", tenantId, rateLimitedRequests);
+                    }
+                });
     }
 }
